feat: Add ActiveMQ configuration and service (main)
This commit introduces ActiveMQ configuration, a
service to interact with it, and exception
handling.  It includes a `ConnectionFactory`
bean and a REST controller to list queues.
ActiveMQ properties are configurable via
`application.yml`. The `application.properties`
file was renamed, and JMX configuration was
added for monitoring.  CORS configuration was
also added.  Swagger configuration was added
for API documentation.  A docker-compose file
was added to simplify setup.